emsw-697: Moved page variable to content stringd
