+++ src/ui/server/content-strings/fields/insurance/your-buyer/index.ts
@@ -1,10 +1,9 @@
-export const yourBuyerPageVariables = () => ({
+export const yourBuyerPageVariables = {
   FIELDS: {
     BUYER_COUNTRY: {
       ID: 'countryId',
-      NAME: 'countryName',
       LABEL: 'Country',
       ERROR_MESSAGE: 'Select your buyer country',
     },
   },
-});
+};
