+++ src/ui/server/controllers/insurance/your-buyer/your-buyer-details/index.ts
@@ -3,9 +3,10 @@ import { Request, Response } from '../../../../../types';
 import api from '../../../../api';
 import { ROUTES, TEMPLATES } from '../../../../constants';
 import insuranceCorePageVariables from '../../../../helpers/page-variables/core/insurance';
-import { yourBuyerPageVariables } from '../../../../constants/field-ids/insurance/your-buyer';
 import { mapCountriesSelect } from '../../../../helpers/mappings/map-countries-select';
 import { yourBuyerDetailsValidation } from './validation/yourBuyerDetailsValidation';
+// import { yourBuyerPageVariables } from '../../../content-strings/fields/insurance/your-buyer';
+import { yourBuyerPageVariables } from '../../../../content-strings/fields/insurance/your-buyer';
 
 export const get = async (req: Request, res: Response) => {
   if (!req.session.submittedData || !req.session.submittedData.insuranceEligibility) {
@@ -28,7 +29,7 @@ export const get = async (req: Request, res: Response) => {
       PAGE_CONTENT_STRINGS: PAGES.INSURANCE.YOUR_BUYER_DETAILS,
       BACK_LINK: req.headers.referer,
     }),
-    ...yourBuyerPageVariables(),
+    ...yourBuyerPageVariables,
     countries: mappedCountries,
   });
 };
@@ -41,7 +42,7 @@ export const post = async (req: Request, res: Response) => {
         PAGE_CONTENT_STRINGS: PAGES.INSURANCE.YOUR_BUYER_DETAILS,
         BACK_LINK: req.headers.referer,
       }),
-      ...yourBuyerPageVariables(),
+      ...yourBuyerPageVariables,
       validationErrors,
     });
   }
